pynodegl: support non iterable arguments in list based parameters
If a named argument is supposed to be a list of elements but the user
only specify one element instead of a list, that single element is
queued in the appropriate list without erroring-out.

The purpose of this commit is to propose a solution to the case where a
list of elements is way less common than a single one. One example of
this is the incoming multi-texture support in the TexturedShape node: a
user should not have to pass a list of 1 texture in the common case.