api: use unique identifiers for node types (API break)
This allows us to reorder, add and delete node identifiers without
breaking the ABI.

Since the mapping can not be kept in a simple array it is now stored in
a macro in a single dedicated header: nodes_register.h.

Unfortunately, GCC nor Clang seem to be able to generate a perfect hash
function based on the u32 identifiers, so the direct mapping with a LUT
ends up becoming successive conditionnal branches when the user creates
a new node. Given the very low amount of nodes (<40), this is not an
issue at all.