camera: adjust eye and center such that the view is an identity by default
This fixes mismatches between preview of scenes without camera and their
export (which includes a camera for the buffer pipe). Mismatch usually
involves clipping.

The view matrix looks like this:

  1  0  0  0        1  0  0  0
  0  1  0  0        0  1  0  0
  0  0  1  0        0  0  1  0
  0  0 -1  1        0  0  0  1

    before            after

Changing the center coordinates instead of the eye prevents a translate
to be included in the matrix, affecting the latest matrix row.