serialize: fix mat4 parameter serialization
With this commit we prevent the serialization of the default matrix.
Serializing the value when it's the default value is not really
problematic, but in our case it was also serializing without the
parameter's key, causing it to be interpreted as a constructor
parameter.

Fixes the following error in the tests:

    [ERROR] nodes.c:770 ngli_node_param_find: parameter 0x1p+0,0x0p+0,0x0p+0,0x0p+0,0x0p+0,0x1p+0,0x0p+0,0x0p+0,0x0p+0,0x0p+0,0x1p+0,0x0p+0,0x0p+0,0x0p+0,0x0p+0,0x1p+0 transform not found in UniformMat4