utils/viewer: stricter mod path blacklisting
This will prevent reloading any Python modules with its root in the
Python system directory.

This commit is motivated by yet another reload problem related to
logging for which the working directory is $sysdir/logging instead of
just $sysdir.

Note that this means $sysdir/site-packages will now be left out. If you
are working in a virtual env or using develop installs of your modules,
this will not be a problem.