ngl-tools: set default OpenGL version to 4.1 on all platforms
Target by default the same OpenGL version that the one given by recent
OSX releases. Good drivers will anyway give us an higher OpenGL version
which is a superset of the one we request on other platforms.

Fixes rendering of scenes using compute shaders as we previously were
requesting an OpenGL version of 3.0 preventing the driver to give us a
context potentially compatible with compute shaders.