serialize/deserialize: use a custom portable representation for floats
In bionic (Android libc), the support for %a was added in printf in
2014, and in scanf in 2016. As a result, it's too early to depend on it,
especially given how the Android ecosystem is maintained.

Some other libc may not have a decent support either. As a result, a
basic exact serialization is defined following the arbitrary "[-]EzM"
format where E and M are respectively the exponent and the fraction
(mantissa) of the float, printed in capitalized hexadecimal (with 32 or
64-bit according to the type being float or double).

In addition to making the float representation exact, this also solves
the locales issue (for instance, printing a float with French locales
will use a comma instead of a dot as separator).