animation: switch to double precision for AnimatedFloat
Single floating point parameters are double (while vectors are floats).
As a result, AnimKeyFrameFloat has a double precision but is
interpolated as float, leading to a loss of precision where it may
matters (typically, times in Media). This commit addresses this issue by
holding the result of the interpolations in the correct storage space.