params: remove unused PARAM_FLAG_USER_SET flag
PARAM_FLAG_USER_SET was introduced for the GraphicConfig node in order
to detect if a given param has been set by the user or not. The usage of
this flag was removed by commit e8fed29464398bd15902b37205ec3db2e8ed34fe
in favor of sensible default param values.