timerangefilter: fix the whole visiting logic
Before this commit, TimeRangeFilter.visit was working on its own state
instead of the child state, in a convoluted way. To address this issue,
it is adjusted to only determine whether its child is supposed to be
active or not, while the generic code is getting back the core code for
handling `is_active` combinations.