api: fix GL context configuration and reconfiguration on iOS
ngli_glcontext_new() calls eagl_init() and eagl_create() which requires
to execute synchronously some code related to the UIView layer on the
main thread. It can lead to a deadlock if the callee of ngl_configure()
is the main thread for non-wrapped context.

This commit adds a custom codepath for ngl_configure() on iOS which
initializes and creates the GL context from the callee thread, makes it
non-current on the callee thread and finally makes it current to the
node.gl worker thread.