camera: remove hflip parameter
The hflip parameter was initially introduced to deal with the OpenGL
coordinate system where the read pixels from the camera framebuffer are
vertically flipped.

The hflip parameter was meant to vertically flip the camera projection
matrix. This is not enough to properly vertically reverse an entire
scene since the camera node transformations do not stack and a graph can
contains cameras at different levels.

The vertical flip of pixels is now done in CPU before pushing it to the
pipe file descriptor.