timerangefilter: fix max idle time condition
Since 6123bfdfd799bac012882d22ce268b24555128e3, the node.state can not
be relied on to check weither a given node is active or not. Indeed, a
node without a prefetch/release defined will be considered STATE_READY
post-init, and not simply initialized. This meant that for every
TimeRangeFilter direct child node without prefetch/release, it was
considered active, and thus all of its children. The effect of this bug
was that in most situation, the actual prefetch value was max_idle_time,
and not prefetch_time.