serial: serialize floats and doubles differently
It is not possible to reconstruct the float or double value if we don't
know the scale information. In the context of node.gl it's not a problem
because we know it from the parameter type, but it is an issue for third
parties trying to parse the serialized form without context.

'z' and 'Z' exponent-mantissa separator characters are respectively used
for floats and doubles.