glcontext_eagl: ensure frame buffers have been initialized before deleting them
Fixes a crash in eagl_uninit() if the EAGL context creation previously
failed in eagl_init() or if one of the functions related to the
glcontext failed in ngli_glcontext_new().

The crash was due to ngli_glDeleteFramebuffers() and
ngli_glDeleteRenderbuffers() trying to use respectively
gl->funcs.DeleteFramebuffers and gl->funcs.DeleteRenderbuffers pointers
that weren't loaded due to the context not being initialized.

Signed-off-by: Matthieu Bouron <matthieu.bouron@gmail.com>