glcontext_eagl: make onscreen initialization only possible from the UI thread
Removes synchronous dispatch calls on the UI thread when called from
another thread.

This commit addresses the issue where the user synchronously dispatches
a call to ngl_configure from the UI thread to another thread. In this
situation, the UI thread is locked until ngl_configure returns. But
since the ngl_configure call has been dispatched on a non-UI thread,
glcontext_eagl.m will try to synchronously dispatch itself back on the
UI thread. Since the UI thread is already waiting synchronously, this
will cause a deadlock.

Adjusting the current pattern such that it doesn't lock would be too
invasive on the public API, so we instead decide to push a constraint on
the user side.