utils/filetracker: simplify file tracking logic
Previously the code was trying to hook every import calls to identify
files. This is actually legacy code from when we were trying to
reconstruct the whole tree in order to reload modules in the appropriate
order. This is not required anymore, so we simply make a diff of
sys.modules before and after the monitored code.

One limitation to keep in mind: it won't detect re-imports of already
imported modules, but in our case it's not important since the usage of
the FileTracker is made in a fresh subprocess.

In addition to simplifying the code, this new logic is compatible with
Python3.