media: clamp remapped time to a minimum of 0
Due to floating point inaccuracies in the time interpolation, the
media_time can end up being slightly negative.

In the current state, it's not possible for the user to trigger a large
negative time aside from abusing the floats inaccuracies thanks to the
monotonic check present in animations. As a result, there is no reason
to keep the current check in the media node.

Following is a scene that can trigger the error:

@scene()
def flat_remap(cfg):
    m0 = cfg.medias[0]
    cfg.duration = m0.duration
    cfg.aspect_ratio = (m0.width, m0.height)

    media_animkf = [
        ngl.AnimKeyFrameFloat(0, 1.833),
        ngl.AnimKeyFrameFloat(cfg.duration, 1.833),
    ]

    q = ngl.Quad((-1, -1, 0), (2, 0, 0), (0, 2, 0))
    m = ngl.Media(m0.filename, time_anim=ngl.AnimatedTime(media_animkf))
    t = ngl.Texture2D(data_src=m)
    render = ngl.Render(q)
    render.update_textures(tex0=t)
    return render