utils/export: rename Exporter.finished() signal
There is already a QThread.finished() signal that is emitted
automatically when the thread exits:
https://doc.qt.io/qt-5/qthread.html#finished

In PyQt5 you could overwrite parent classes signals. With PySide2 the
existing QThread.finished() is still triggered, even though we could
think it was overwritten by Exporter.finished(), leading to connected
slots being called twice.

In the case of Exporter.finished(), the signal must only be emitted if
the export actually succeeded. Hence the chosen solution here is to
rename the signal to Exporter.export_finished().