shaders: use main() instead of main(void)
We have about 30 matches of each in the codebase, this commit makes it
consistent. We choose the "main()" form as it is the one advertized in
the GLSL specs. While it makes a difference in C, it doesn't in GLSL as
the function can not take any argument anyway. Last but not least, the
glslang repository has currently 600+ matches for "main()" and 2 for
"main(void)".