api: remove va-arg constructor mechanism
va-arg are clumsy and error prone, mainly due to the abscence of type
checking.

This commit makes the usage of nodes with constructor unsafe temporarily
(because some parameter can now be NULL); the relevant runtime checks
are introduced in a future commit.

Note: the binding gets bigger with this commit: from 7.5M on master
(915K stripped) to 8.2M (994K stripped), due to the number of added
explicit parameter inits in the constructor (which were previously
inlined as va-arg).