tools/player: reset running time after a seek request
The first call to ngl_draw() can take quite some time as it initializes
and prefetch all the resources required for the graph for a given time.
Calling ngl_draw() after a seek request can also take quite some time as
it needs to release/prefetch all the requires resources for a given
time.