log: force stdout flushing after logging a line
This addresses a flushing issue at least on MinGW where both stderr and
stdout are fully buffered, even when using fprintf.  An alternative
would be to configure the flushing policy with setvbuf(), but this can
have unexpected behaviour for library users. Typically, we don't want to
get unbuffered behaviour in the ngl-tools where the user make use of the
pipes to communicate data between tools.