utils/graph_view: remove playback mechanism
While it makes sense to observe the resources of a tree at a given time,
having a playback that run a subprocess 60 times per second is not
exactly practicable. The feature was present because the feature was
shared with the widget player, but the complexity added by this real
time system is not worth the maintenance burden.