rtt: remove vflip parameter
This parameter was useful for users relying on the OpenGL backend and
prepared the geometry and uv coordinates to deal with the rendertarget
output being vertically flipped.

Since df18098f01ff9795018b23dfc844225f325db527, we alter the
rendertarget output coordinate matrix for OpenGL offscreen rendering,
leaving this parameter not working as expected in this case.

Moreover, this parameter won't work as expected with other graphics API.