tools/common: define _POSIX_C_SOURCE for clock_gettime()
This is needed for clock_gettime() and its associated timespec struct.
The reason it currently works is because of -D_REENTRANT being added by
the SDL C flags, which implies a recent _POSIX_C_SOURCE.

Not all tools are built with SDL, but we are lucky the first one
compiling common.c does it with the SDL flags. This depends on the
undefined behaviour of evaluation of the prerequisite orders, and we
should not rely on that.