build: move GL backend specific code to a dedicated directory
This operation is a simple move to help the isolation of the backends
themselves.

Files have not been renamed, nor the include directives adjusted because
are automatically picked relatively to the current directory of the
source file. This also means that removing the _gl suffixes from
filenames would typically still clash with the identically named files
in libnodegl root directory. This needs to be addressed carefully in a
future commit.

The only source modifications are in api.c and node_media.c,
respectively for Apple and Android platforms.