internal: remove gtimer abstraction
The current gtimer abstraction is not compatible with modern backends as
they require an explicit command buffer to write the timestamp queries.
Command buffers are not exposed by the current node.gl graphics
abstraction. Doing so would be quite complex because we would need to
emulate the entire command buffer API for the OpenGL backends.