tools/build: rely on meson to check if tool_deps can be honored
Relies on meson to check if tool_deps can be honored instead of manually
checking if all the dependencies can be found. This allows non-mandatory
dependencies that could not be found to not disable an entire target.

Fixes targets availability on system that do not provide a separate math
library (MSVC for example).