api: do not call cmd_stop() on context initialization failure
If the graphics context and/or its resources fail to initialize (either
in cmd_configure() or cmd_set_capture_buffer()) the node.gl context is
marked as unconfigured. This allows other API calls that depend on
theses resources to fail early on.

We do not need to release the partially initialized resources if an
error happens on context initialization as they will be released with
the next call to ngl_configure() or ngl_freep().