nodes: invalidate the node and its parents on live change
The commit makes the nodes invalidate themselves as well as their
parents (up to the root node) and makes them candidates for update in
the next call to ngl_draw(). This allows live changes to have effect
even if the draw time does not change (this typically happens when a
player is paused).

A new node callback, invalidate, has been added to allow a node
implementation to perform specific actions when a live change happens
in the node itself or in one node of its children.

For simple live change cases where there is not data dependency between
the node and its parents (Camera, Scale, Rotate, RotateQuat, Translate,
Transform, Skew, ...), this commit will allow to remove the need to
compute any data directly in the live change callback as we are
guaranteed that the node and its parents will be updated in the next
ngl_draw() call.

This commits will also allow to fix issues when live changes happen to a
node used a dependency by another node (Uniforms attached to a Block
for example).