tests: add tolerance in rtt_clear_attachment_with_timeranges
Some frames have a few off-by-one pixels in the green component which
actually cascade down to the timber sized (9x9) picture of the finger print.
The reason for that is probably because these off-by-one are happening on 0x7F
vs 0x80 values (the middle of the 8-bit range). These offsets on the border of
the shape cause a slightly discontinuity in the location of the sign
differences in the final fingerprint.

The visual difference is unoticeable and perfectly acceptable, so a minimal 3%
tolerance is chosen to address the problem.

Reproducible on GeForce GTX 1050 Ti Mobile (with official drivers 450.119.03).